API

API for TaylorDiff.

TaylorDiff.TaylorScalar — Type

TaylorScalar{T, N}

Representation of Taylor polynomials.

Fields

value::NTuple{N, T}: i-th element of this stores the (i-1)-th derivative

TaylorDiff.TaylorScalar — Method

TaylorScalar{T, N}(x::T, d::T) where {T, N}

Construct a Taylor polynomial with zeroth and first order coefficient, acting as a seed.

TaylorDiff.TaylorScalar — Method

TaylorScalar{T, N}(x::T) where {T, N}

Construct a Taylor polynomial with zeroth order coefficient.

TaylorDiff.can_taylorize — Method

TaylorDiff.can_taylor(V::Type)

Determines whether the type V is allowed as the scalar type in a Dual. By default, only <:Real types are allowed.

TaylorDiff.derivative — Function

derivative(f, x, l, ::Val{N})
derivative(f!, y, x, l, ::Val{N})

Computes order-th directional derivative of f w.r.t. vector x in direction l.

TaylorDiff.derivative! — Function

derivative!(result, f, x, l, ::Val{N})
derivative!(result, f!, y, x, l, ::Val{N})

In-place derivative calculation APIs. result is expected to be pre-allocated and have the same shape as y.

TaylorDiff.derivatives — Function

derivatives(f, x, l, ::Val{N})
derivatives(f!, y, x, l, ::Val{N})

Computes all derivatives of f at x up to order N - 1.